services/ingestion/requirements.txt
pdfminer.six>=20221105
python-docx>=1.0.0
pandas>=2.0.0
openpyxl>=3.1.0
pytest>=8.0.0
